+++ src/http.h
@@ -12,12 +12,16 @@ namespace taranis {
 
 class HttpClient {
 public:
+  std::string encode_query_parameter(const std::string &parameter) const;
+
   Json::Value get(const std::string &url);
 
 private:
-  std::unique_ptr<CURL, void (*)(CURL *)> preprare_curl();
+  std::string response_data;
 
-  void ensure_network();
+  std::unique_ptr<CURL, void (*)(CURL *)> prepare_curl();
+
+  void ensure_network() const;
 
   static size_t write_callback(void *contents, size_t size, size_t nmemb,
                                void *userp);